requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-multipart==0.0.6